// src/BookBarter.React/src/api/generated/models/update-book-command.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BookBarter.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UpdateBookCommand
 */
export interface UpdateBookCommand {
    /**
     * 
     * @type {number}
     * @memberof UpdateBookCommand
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof UpdateBookCommand
     */
    'isbn'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateBookCommand
     */
    'title'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateBookCommand
     */
    'publicationDate'?: string;
    /**
     * 
     * @type {Array<number>}
     * @memberof UpdateBookCommand
     */
    'authorsIds'?: Array<number> | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateBookCommand
     */
    'genreId'?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateBookCommand
     */
    'publisherId'?: number;
}